Consolidate alter tests in context tests
